+++ src/refactoring/hooks/useAdmin.ts
@@ -1,13 +1,7 @@
 import { useState } from "react";
-import { Coupon, Discount, Product } from "../../types";
+import { Coupon, Product } from "../../types";
 
 const useAdmin = () => {
-  const [openProductIds, setOpenProductIds] = useState<Set<string>>(new Set());
-  const [editingProduct, setEditingProduct] = useState<Product | null>(null);
-  const [newDiscount, setNewDiscount] = useState<Discount>({
-    quantity: 0,
-    rate: 0,
-  });
   const [newCoupon, setNewCoupon] = useState<Coupon>({
     name: "",
     code: "",
@@ -21,97 +15,6 @@ const useAdmin = () => {
     stock: 0,
     discounts: [],
   });
-
-  const toggleProductAccordion = (productId: string) => {
-    setOpenProductIds((prev) => {
-      const newSet = new Set(prev);
-      if (newSet.has(productId)) {
-        newSet.delete(productId);
-      } else {
-        newSet.add(productId);
-      }
-      return newSet;
-    });
-  };
-
-  // handleEditProduct 함수 수정
-  const handleEditProduct = (product: Product) => {
-    setEditingProduct({ ...product });
-  };
-
-  // 새로운 핸들러 함수 추가
-  const handleProductNameUpdate = (productId: string, newName: string) => {
-    if (editingProduct && editingProduct.id === productId) {
-      const updatedProduct = { ...editingProduct, name: newName };
-      setEditingProduct(updatedProduct);
-    }
-  };
-
-  // 새로운 핸들러 함수 추가
-  const handlePriceUpdate = (productId: string, newPrice: number) => {
-    if (editingProduct && editingProduct.id === productId) {
-      const updatedProduct = { ...editingProduct, price: newPrice };
-      setEditingProduct(updatedProduct);
-    }
-  };
-
-  // 수정 완료 핸들러 함수 추가
-  const handleEditComplete = (
-    onProductUpdate: (updatedProduct: Product) => void
-  ) => {
-    if (editingProduct) {
-      onProductUpdate(editingProduct);
-      setEditingProduct(null);
-    }
-  };
-
-  const handleStockUpdate = (
-    productId: string,
-    newStock: number,
-    products: Product[],
-    onProductUpdate: (updatedProduct: Product) => void
-  ) => {
-    const updatedProduct = products.find((p) => p.id === productId);
-    if (updatedProduct) {
-      const newProduct = { ...updatedProduct, stock: newStock };
-      onProductUpdate(newProduct);
-      setEditingProduct(newProduct);
-    }
-  };
-
-  const handleAddDiscount = (
-    productId: string,
-    products: Product[],
-    onProductUpdate: (updatedProduct: Product) => void
-  ) => {
-    const updatedProduct = products.find((p) => p.id === productId);
-    if (updatedProduct && editingProduct) {
-      const newProduct = {
-        ...updatedProduct,
-        discounts: [...updatedProduct.discounts, newDiscount],
-      };
-      onProductUpdate(newProduct);
-      setEditingProduct(newProduct);
-      setNewDiscount({ quantity: 0, rate: 0 });
-    }
-  };
-
-  const handleRemoveDiscount = (
-    productId: string,
-    index: number,
-    products: Product[],
-    onProductUpdate: (updatedProduct: Product) => void
-  ) => {
-    const updatedProduct = products.find((p) => p.id === productId);
-    if (updatedProduct) {
-      const newProduct = {
-        ...updatedProduct,
-        discounts: updatedProduct.discounts.filter((_, i) => i !== index),
-      };
-      onProductUpdate(newProduct);
-      setEditingProduct(newProduct);
-    }
-  };
 
   const handleAddCoupon = (onCouponAdd: (newCoupon: Coupon) => void) => {
     onCouponAdd(newCoupon);
@@ -141,18 +44,6 @@ const useAdmin = () => {
     newProduct,
     setNewProduct,
     handleAddNewProduct,
-    toggleProductAccordion,
-    openProductIds,
-    editingProduct,
-    handleProductNameUpdate,
-    handlePriceUpdate,
-    handleStockUpdate,
-    handleRemoveDiscount,
-    newDiscount,
-    setNewDiscount,
-    handleAddDiscount,
-    handleEditComplete,
-    handleEditProduct,
     newCoupon,
     setNewCoupon,
     handleAddCoupon,
