+++ src/refactoring/utils/formatUtils.ts
@@ -1,5 +1,9 @@
 import { Coupon } from "../../types";
 
+const formatDiscount = (discountValue: number, discountType: string) => {
+  return `${discountValue}${formatDiscountType(discountType)}`;
+};
+
 const formatDiscountType = (discountType: string) => {
   if (discountType === "amount") return "원";
   return "%";
@@ -7,16 +11,25 @@ const formatDiscountType = (discountType: string) => {
 
 export const formatCoupon = (coupon: Coupon) => {
   const { name, discountType, discountValue } = coupon;
-  return `${name} - ${discountValue}${formatDiscountType(discountType)}`;
+  return `${name} - ${formatDiscount(discountValue, discountType)}`;
 };
 
 export const formatSelectedCoupon = (coupon: Coupon) => {
   const { name, discountType, discountValue } = coupon;
-  return `적용된 쿠폰: ${name}(${discountValue}${formatDiscountType(
+  return `적용된 쿠폰: ${name}(${formatDiscount(
+    discountValue,
     discountType
-  )} 할인)`;
+  )}} 할인)`;
 };
 
 export const formatCommaWithUnit = (amount: number) => {
   return `${amount.toLocaleString()}원`;
+};
+
+export const formatAdminCoupon = (coupon: Coupon) => {
+  const { name, code, discountType, discountValue } = coupon;
+  return `${name} (${code}):${formatDiscount(
+    discountValue,
+    discountType
+  )} 할인`;
 };
